Move interval monitoring to the video component rather than play component
