+++ src/pages/film-room/[video]/index.tsx
@@ -30,8 +30,11 @@ const FilmRoom = () => {
   const [player, setPlayer] = useState<YouTubePlayer | null>(null);
   const playerRef = useRef<HTMLDivElement | null>(null);
   const [activePlay, setActivePlay] = useState<PlayPreviewType | null>(null);
+  const [seenActivePlay, setSeenActivePlay] = useState<boolean>(false);
 
   const [isNewPlayOpen, setIsNewPlayOpen] = useState<boolean>(false);
+
+  const interval = useRef<NodeJS.Timeout | null>(null);
 
   const fetchVideo = async () => {
     const { data } = await supabase
@@ -51,7 +54,7 @@ const FilmRoom = () => {
   };
 
   const fetchActivePlay = async () => {
-    if (playParam && user.isLoggedIn) {
+    if (playParam) {
       const { data } = await supabase
         .from("play_preview")
         .select(`*`, {
@@ -60,6 +63,7 @@ const FilmRoom = () => {
         .eq("play->>id", playParam)
         .single();
       if (data) setActivePlay(data);
+      else setActivePlay(null);
     }
   };
 
@@ -69,6 +73,7 @@ const FilmRoom = () => {
     const time = Number(startParam);
     if (time) {
       void video.seekTo(time, true);
+      void video.playVideo();
     }
   };
 
@@ -76,18 +81,68 @@ const FilmRoom = () => {
     if (playerRef) playerRef.current?.scrollIntoView({ behavior: "smooth" });
   };
 
+  const getCurrentTime = async (player: YouTubePlayer) => {
+    return Math.round(await player.getCurrentTime());
+  };
+
+  const checkTime = async () => {
+    if (player && activePlay) {
+      void getCurrentTime(player).then((currentTime) => {
+        const endTime = activePlay.play.end_time;
+        if (currentTime == endTime && interval.current && !seenActivePlay) {
+          void player.pauseVideo();
+          void clearInterval(interval.current);
+          interval.current = null;
+          setSeenActivePlay(true);
+        } else if (currentTime !== endTime && !interval.current) {
+          interval.current = setInterval(async () => void checkTime(), 1000);
+        } else if (currentTime == endTime && seenActivePlay) {
+          void player.seekTo(activePlay.play.start_time, true);
+          void player.pauseVideo();
+          if (interval.current) {
+            void clearInterval(interval.current);
+            interval.current = null;
+          }
+        }
+        return;
+      });
+    } else return;
+  };
+
   useEffect(() => {
     if (router.query.video) {
       void fetchVideo();
       void fetchAffiliatedTeams();
-      void fetchActivePlay();
     }
-  }, [router.query.video, playParam]);
+  }, [router.query.video]);
+
+  useEffect(() => {
+    void fetchActivePlay();
+  }, [playParam, player]);
+
+  useEffect(() => {
+    if (interval.current) {
+      clearInterval(interval.current);
+      interval.current = null;
+    }
+    if (activePlay && player) {
+      void checkTime();
+    }
+  }, [activePlay, seenActivePlay]);
 
   useEffect(() => {
     const time = Number(startParam);
     if (time && player) void player.seekTo(time, true);
   }, [startParam]);
+
+  useEffect(() => {
+    return () => {
+      if (interval.current) {
+        clearInterval(interval.current);
+        interval.current = null;
+      }
+    };
+  }, []);
 
   return (
     video && (
@@ -146,6 +201,7 @@ const FilmRoom = () => {
           scrollToPlayer={scrollToPlayer}
           setActivePlay={setActivePlay}
           activePlay={activePlay}
+          setSeenActivePlay={setSeenActivePlay}
         />
       </div>
     )
